Another better progress formula.
The last one would choose quality over all else, because it only
down-weighted CP with the successProbability factor instead of the
entire synth.  This change weights the entire formula according to
probability of success, and also adds an additional unit test to
ensure that all other things being equal, when the success probability
decreases, so does the score.