Reverted previous "de-optimization", and fixed some UI issues.
I decided that, although the de-optimization is technically correct,
it is not worth the tradeoff.  Conceptually, all it does is to
individually descend into states that differ only by condition
separately.  While theoretically this is optimal, it increases the
branching factor substantially, leading to exponentially longer
analysis times, and only improves analysis quality slightly.  Without
this fix, one can increase analysis quality by increasing the depth,
and get far superior results in a fraction of the time.

Also changed the color of disabled radio buttons to be more distinct
from enabled radio buttons, and fixed an issue where radio buttons
corresponding to an impossible random outcome are sometimes enabled.